Move index-view-model-factoy to seperate namespace.
