+++ src/Example.Web/Controllers/HomeController.cs
@@ -1,4 +1,5 @@
-﻿using Example.Web.ViewModels.Home;
+﻿using Example.Web.Factories.ViewModels.Home;
+using Example.Web.ViewModels.Home;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Example.Web.Controllers;
@@ -14,7 +15,10 @@ public class HomeController : Controller
     [IgnoreAntiforgeryToken] // yes, this should not be there, but for this simple example it makes our integration testing much easier
     public async Task<IActionResult> Index([FromServices] IndexViewModelFactory factory, IndexViewModel viewModel)
     {
-        if (!ModelState.IsValid) return View(viewModel);
+        if (!ModelState.IsValid)
+        {
+            return View(viewModel);
+        }
         return View(await factory.HandleRegistration(viewModel));
     }
 }
